+++ frontend/app/studio/[brainId]/BrainManagementTabs/components/KnowledgeTab/components/AddedKnowledge/AddedKnowledge.tsx
