+++ frontend/app/studio/[brainId]/BrainManagementTabs/components/KnowledgeTab/components/KnowledgeTable/KnowledgeTable.tsx
