+++ frontend/app/studio/[brainId]/BrainManagementTabs/components/KnowledgeTab/components/KnowledgeTable/components/KnowledgeItem/components/DeleteKnowledge.tsx
