+++ frontend/app/studio/[brainId]/BrainManagementTabs/components/KnowledgeTab/components/KnowledgeTable/components/KnowledgeItem/hooks/useKnowledgeItem.ts
