+++ frontend/app/studio/[brainId]/BrainManagementTabs/components/KnowledgeTab/components/KnowledgeTable/components/KnowledgeItem/index.tsx
